Remove debug lines

+++ include/ford_mot/utilities.h
@@ -18,7 +18,7 @@ class Utilities
         void printHypothesis(std::vector<std::vector<float>> gij);
 
         std::vector<std::vector<int>> hypothesisCombinations(std::vector<std::vector<float>> gijList);
-        std::vector<std::unordered_set<int>> hypothesisMerge(std::vector<std::unordered_set<int>> combinations, std::vector<float> trackObservationList);
+        std::vector<std::unordered_set<int>> hypothesisMerge(std::vector<std::unordered_set<int>> combinations, std::vector<int> trackObservationList);
 
         // bool sortbysec(const std::pair<int,int> &a,const std::pair<int,int> &b);
